+++ projects/tuist/fixtures/ios_app_with_xcframeworks/Modules/StaticFrameworkA/Sources/StaticFrameworkA.swift
@@ -1,13 +1,15 @@
 import Foundation
 import MyStaticLibrary
+import StaticFrameworkB
 
 public final class StaticFrameworkAComponent {
     public let name = "StaticFrameworkAComponent"
     private let staticLibraryComponent = MyStaticLibraryComponent()
+    private let staticFrameworkBComponent = StaticFrameworkBComponent()
 
     public init() {}
 
     public func composedName() -> String {
-        "\(name) > \(staticLibraryComponent.name)"
+        "\(name) > \(staticLibraryComponent.name) > \(staticFrameworkBComponent.composedName())"
     }
 }